requirements.txt
# Dependencias para el proyecto de recuperación de imágenes

# Deep Learning y CNNs
tensorflow>=2.13.0
keras>=2.13.0

# Procesamiento de imágenes
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Análisis numérico
numpy>=1.24.0
scipy>=1.11.0

# Búsqueda y similitud
faiss-cpu>=1.7.4  # Usar faiss-gpu si tienes GPU con CUDA

# Utilidades
tqdm>=4.66.0
matplotlib>=3.7.0

# Opcional: para visualización avanzada
seaborn>=0.12.0
